Port CoreTest and DispatchLoopTest to CXXTest.
