+++ test/zcm/wscript
@@ -7,9 +7,3 @@ def build(ctx):
               source = 'multi_file.cpp',
               rpath = ctx.env.RPATH_zcm,
               install_path = None)
-
-    ctx.program(target = 'dispatch_loop',
-                use = 'default zcm',
-                source = 'dispatch_loop.cpp',
-                rpath = ctx.env.RPATH_zcm,
-                install_path = None)
